fe: Refactor App.jsx by moving resource data fetching to new `useResourceData` hook
